refactor : formatDate함수 util 분리 및 Invitation 타입 전역 선언
- 서버에서 받아오는 초대장 데이터 형식인 Invitation 타입을 전역타입으로 분리하였습니다.
- locale 데이터를 YYYY.MM.DD로 변환하는 formatDate 함수를 util함수로 분리하였습니다.

ref : #28
