+++ src/utils/formatDate.ts
@@ -1,0 +1,10 @@
+// locale 데이터 => YYYY.MM.DD 변환 함수
+export const formatDate = (dateString: string): string => {
+  const date = new Date(dateString);
+
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  const day = date.getDate();
+
+  return `${year}년 ${month}월 ${day}일`;
+};
